fix and update

+++ zcrypto/aes.c
@@ -62,10 +62,10 @@ static inline uint32_t _sub_word(uint32_t x) {
     return *(uint32_t*)y;
 }
 
-static void aes_set_key(const uint8_t *key, int keylen, int round, uint32_t *rkey) {
+static void aes_set_key(const uint8_t *key, size_t keylen, int round, uint32_t *rkey) {
     memcpy(rkey, key, keylen);
-    int kn = keylen / 4;
-    for (int r = kn; r < (round + 1) * 4; ++r) {
+    size_t kn = keylen / 4;
+    for (size_t r = kn; r < (size_t)(round + 1) * 4; ++r) {
         if (r % kn == 0) {
             uint32_t rcon = 0;
             *(uint8_t*)&rcon = RCBOX[r / kn];
